requirements.txt
django-admin-thumbnails==0.2.8
django-cors-headers==4.3.1
django-filter==23.4
djangorestframework==3.14.0
drf-writable-nested==0.7.0
pillow==10.1.0
pip-chill==1.0.3
